Remove (mostly) dead code and refactor the few dependencies that exist
