+++ src/features/views/components/ViewDataTable/columnTypes/SurveyResponseColumnType.tsx
@@ -11,7 +11,7 @@ import SurveySubmissionPane from 'features/surveys/panes/SurveySubmissionPane';
 import { usePanes } from 'utils/panes';
 import ViewSurveySubmissionPreview from '../../ViewSurveySubmissionPreview';
 
-type SurveyResponseViewCell = {
+export type SurveyResponseViewCell = {
   submission_id: number;
   submitted: string;
   text: string;
